move easydocs into com.github.kfang package
