+++ jvm/src/main/scala/com/github/kfang/easydocs/AppPackage.scala
@@ -1,0 +1,9 @@
+package com.github.kfang.easydocs
+
+import akka.actor.ActorSystem
+
+case class AppPackage(
+  system: ActorSystem,
+  config: AppConfig,
+  services: AppServices
+)
