+++ jvm/src/main/scala/com/github/kfang/easydocs/routes/responses/Response.scala
@@ -1,0 +1,30 @@
+package com.github.kfang.easydocs.routes.responses
+
+import com.github.kfang.easydocs.utils.ERR
+import spray.json.JsObject
+import spray.routing.Route
+import spray.routing.directives.FutureDirectives._
+import spray.routing.directives.RouteDirectives._
+import scala.concurrent.{Future, ExecutionContext}
+import scala.language.implicitConversions
+import scala.util.{Failure, Success}
+import spray.json.DefaultJsonProtocol._
+import spray.httpx.SprayJsonSupport._
+
+trait Response {
+
+  implicit def jsObjectToRoute(js: JsObject): Route = complete(js)
+
+  implicit def futureJsObjectToRoute(fjs: Future[JsObject])(implicit ec: ExecutionContext): Route = onComplete(fjs){
+    case Success(js) => complete(js)
+    case Failure(e)  => ERR(e).complete
+  }
+
+  implicit def futureString(s: Future[String])(implicit ec: ExecutionContext): Route = onComplete(s){
+    case Success(str) => complete(str)
+    case Failure(e) => ERR(e).complete
+  }
+
+  def finish(implicit ec: ExecutionContext): Route
+
+}
