--- jvm/src/main/scala/easydocs/AppPackage.scala
@@ -1,9 +1,0 @@
-package easydocs
-
-import akka.actor.ActorSystem
-
-case class AppPackage(
-  system: ActorSystem,
-  config: AppConfig,
-  services: AppServices
-)
