--- jvm/src/main/scala/easydocs/routes/ExportRoutes.scala
@@ -1,20 +1,0 @@
-package easydocs.routes
-
-import com.sksamuel.elastic4s.ElasticClient
-import easydocs.AppPackage
-import easydocs.routes.responses.{ExportEndpointsResponse, ExportSitesResponse}
-
-class ExportRoutes(implicit App: AppPackage) extends ExtendedDirectives(App){
-  import App.system.dispatcher
-  implicit private val elasticClient: ElasticClient = App.services.elasticClient
-  
-  
-  private val exportSites = (get & path("sites")){ ExportSitesResponse(elasticClient) }
-  private val exportEndpoints = (get & path("endpoints")){ ExportEndpointsResponse(elasticClient) }
-
-  val routes= pathPrefix("export"){
-    exportSites ~
-    exportEndpoints
-  }
-
-}
